phys improvements
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@2096 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ classes/protocols/stimManagers/@ifFeatureGoRightWithTwoFlank/isTargetFlankerOn.m
@@ -5,7 +5,7 @@ function   [targetIsOn flankerIsOn effectiveFrame cycleNum sweptID repetition]=i
 
 if ~isempty(s.dynamicSweep) 
     %determine the effective frame
-    framesPerCycle=max([s.targetOnOff s.flankerOnOff]);
+    framesPerCycle=double(max([s.targetOnOff s.flankerOnOff]));
     cycleNum=floor(frame/framesPerCycle)+1;
     effectiveFrame=mod(frame, framesPerCycle);
